+++ src/main/java/imhong/dowith/challenge/repository/ChallengeRepository.java
@@ -1,5 +1,6 @@
-package imhong.dowith.challenge.domain;
+package imhong.dowith.challenge.repository;
 
+import imhong.dowith.challenge.domain.Challenge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
