+++ src/main/java/imhong/dowith/challenge/repository/ImageRepository.java
@@ -1,5 +1,6 @@
-package imhong.dowith.challenge.domain;
+package imhong.dowith.challenge.repository;
 
+import imhong.dowith.challenge.domain.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
