+++ src/main/java/imhong/dowith/challenge/repository/MemberChallengeRepository.java
@@ -1,5 +1,6 @@
-package imhong.dowith.challenge.domain;
+package imhong.dowith.challenge.repository;
 
+import imhong.dowith.challenge.domain.MemberChallenge;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
