+++ src/main/java/imhong/dowith/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package imhong.dowith.member.domain;
+package imhong.dowith.member.repository;
 
+import imhong.dowith.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
