+++ src/test/java/imhong/dowith/challenge/controller/ChallengeControllerTest.java
@@ -5,16 +5,16 @@ import static org.springframework.http.HttpHeaders.LOCATION;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import imhong.dowith.challenge.domain.Challenge;
-import imhong.dowith.challenge.domain.ChallengeRepository;
 import imhong.dowith.challenge.domain.ChallengeStatus;
 import imhong.dowith.challenge.domain.MemberChallenge;
-import imhong.dowith.challenge.domain.MemberChallengeRepository;
 import imhong.dowith.challenge.dto.ChallengeCreateRequest;
+import imhong.dowith.challenge.repository.ChallengeRepository;
+import imhong.dowith.challenge.repository.MemberChallengeRepository;
 import imhong.dowith.common.ControllerTest;
 import imhong.dowith.common.ImageFileGenerator;
 import imhong.dowith.common.ResponseHeaderHelper;
 import imhong.dowith.member.domain.Member;
-import imhong.dowith.member.domain.MemberRepository;
+import imhong.dowith.member.repository.MemberRepository;
 import io.restassured.response.Response;
 import java.io.IOException;
 import java.time.LocalDate;
